backend/requirements.txt
flask
flask-cors
opencv-python-headless
numpy
paddleocr
easyocr
Pillow
python-multipart
